+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from "./contact-details/index.js";
 export * from "./footer/index.js";
 export * from "./context-menu/index.js";
 export * from "./user-profile/index.js";
+export * from "./confirmation-dialog/index.js";